graphql/backend/requirements.txt
Django>=4.1,<4.2
wagtail>=4.2,<4.3
gunicorn==20.1.0
unicorn==2.0.1.post1
uvicorn==0.21.1
wagtailcodeblock==1.29.0.0
psycopg2-binary==2.9.5
channels==4.0.0
djangorestframework==3.14.0
graphene-django==3.0.0
daphne==4.0.0
ipython==8.12.0
channels_redis==4.1.0
selenium==4.9.1
celery==5.2.7
faker==18.9.0
graphene-django==3.0.0
